feat[frontend]: add theme
