+++ frontend/hero-theme.ts
@@ -1,0 +1,16 @@
+import { heroui } from "@heroui/react"
+
+export default heroui({
+  themes: {
+    dark: {
+      colors: {
+        background: "#111111",
+        primary: {
+          DEFAULT: "#BEF264",
+          foreground: "#111111",
+        },
+        focus: "#BEF264",
+      },
+    },
+  },
+})
