--- frontend/hero.ts
@@ -1,2 +1,0 @@
-import { heroui } from "@heroui/react"
-export default heroui()
