+++ frontend/style.css
@@ -1,6 +1,6 @@
 @import "tailwindcss";
 
-@plugin "./hero.ts";
+@plugin "./hero-theme.ts";
 @source '../node_modules/@heroui/theme/dist/**/*.{js,ts,jsx,tsx}';
 @custom-variant dark (&:where(.dark, .dark *));
